+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * Flowbite::Link component to render links.
+* Flowbite::Card now displays a title via the title argument/slot.
 
 ### Changed